Инициализировал карту
